requirements.txt
# Core trading stack
solana
solders
anchorpy
base58
pumpfun-sdk
gmgn-python-sdk
jupiter-api
dexscreener-api
birdeye-so

# Web3 and blockchain
web3
eth-account
eth-typing
moralis
pycoingecko
ccxt

# Data processing
pandas
numpy
scipy
scikit-learn
ta
yfinance
plotly
dash
streamlit

# Async and networking
aiohttp
websockets
asyncio
requests
httpx
redis
celery

# Telegram bot
python-telegram-bot
telethon
aiogram

# Web framework
fastapi
uvicorn
sqlalchemy
alembic
pydantic
pydantic-settings

# Monitoring and logging
prometheus-client
grafana-client
structlog
python-json-logger
sentry-sdk

# Additional utilities
python-dotenv
pyyaml
toml
cryptography
psycopg2-binary
redis
apscheduler
tenacity
backoff
